move and rename MockLookupProvider
