+++ packages/components/src/test/MockLookupProvider.tsx
@@ -1,15 +1,15 @@
 import React, { Component, PropsWithChildren } from 'react';
 import { List } from 'immutable';
 
-import { ILookupContext, LookupContextProvider } from '../../internal/components/domainproperties/Lookup/Context';
-import { handleSchemas, processContainers, processQueries } from '../../internal/components/domainproperties/actions';
-import { QueryInfoLite } from '../../internal/components/domainproperties/models';
+import { ILookupContext, LookupContextProvider } from '../internal/components/domainproperties/Lookup/Context';
+import { handleSchemas, processContainers, processQueries } from '../internal/components/domainproperties/actions';
+import { QueryInfoLite } from '../internal/components/domainproperties/models';
 
-import containerData from '../data/project-getContainers.json';
-import queryData from '../data/query-getQueries.json';
-import schemaData from '../data/query-getSchemas.json';
-import { Container } from '../../internal/components/base/models/Container';
-import { SchemaDetails } from '../../internal/SchemaDetails';
+import containerData from './data/project-getContainers.json';
+import queryData from './data/query-getQueries.json';
+import schemaData from './data/query-getSchemas.json';
+import { Container } from '../internal/components/base/models/Container';
+import { SchemaDetails } from '../internal/SchemaDetails';
 
 export class MockLookupProvider extends Component<PropsWithChildren, ILookupContext> {
     constructor(props) {
